+++ src/main/gen/io/runescript/plugin/lang/psi/RsPrefixExpression.java
@@ -1,0 +1,16 @@
+// This class is automatically generated. Do not edit.
+package io.runescript.plugin.lang.psi;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.psi.PsiElement;
+
+public interface RsPrefixExpression extends RsExpression, RsUnaryExpression {
+
+  @NotNull
+  RsExpression getExpression();
+
+  @NotNull
+  RsUnaryOp getUnaryOp();
+
+}