+++ src/main/gen/io/runescript/plugin/lang/psi/RsUnaryOp.java
@@ -1,0 +1,10 @@
+// This class is automatically generated. Do not edit.
+package io.runescript.plugin.lang.psi;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.psi.PsiElement;
+
+public interface RsUnaryOp extends PsiElement {
+
+}