+++ src/main/gen/io/runescript/plugin/lang/psi/impl/RsUnaryOpImpl.java
@@ -1,0 +1,30 @@
+// This class is automatically generated. Do not edit.
+package io.runescript.plugin.lang.psi.impl;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.lang.ASTNode;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiElementVisitor;
+import com.intellij.psi.util.PsiTreeUtil;
+import static io.runescript.plugin.lang.psi.RsElementTypes.*;
+import com.intellij.extapi.psi.ASTWrapperPsiElement;
+import io.runescript.plugin.lang.psi.*;
+
+public class RsUnaryOpImpl extends ASTWrapperPsiElement implements RsUnaryOp {
+
+  public RsUnaryOpImpl(@NotNull ASTNode node) {
+    super(node);
+  }
+
+  public void accept(@NotNull RsVisitor visitor) {
+    visitor.visitUnaryOp(this);
+  }
+
+  @Override
+  public void accept(@NotNull PsiElementVisitor visitor) {
+    if (visitor instanceof RsVisitor) accept((RsVisitor)visitor);
+    else super.accept(visitor);
+  }
+
+}