+++ src/main/kotlin/io/runescript/plugin/lang/psi/RsBinaryExpression.kt
@@ -1,8 +1,5 @@
 package io.runescript.plugin.lang.psi
 
-/**
- * test
- */
 interface RsBinaryExpression : RsExpression {
     val left: RsExpression
     val right: RsExpression