+++ src/main/kotlin/io/runescript/plugin/lang/psi/RsUnaryExpression.kt
@@ -1,0 +1,6 @@
+package io.runescript.plugin.lang.psi
+
+interface RsUnaryExpression : RsExpression {
+    val expression: RsExpression
+    val unaryOp: RsUnaryOp
+}